Reorganize importing QKeras (part 1)
