+++ hls4ml/model/optimizer/__init__.py
@@ -26,44 +26,30 @@ optimizers = extract_optimizers_from_path(opt_path, module_path)
 for opt_name, opt in optimizers.items():
     register_pass(opt_name, opt)
 
-try:
-    import qkeras  # noqa: F401
-
-    register_flow(
-        'convert',
-        [
-            'fuse_bias_add',
-            'remove_useless_transpose',
-            'output_rounding_saturation_mode',
-            'qkeras_factorize_alpha',
-            'extract_ternary_threshold',
-            'fuse_consecutive_batch_normalization',
-        ],
-    )  # TODO Maybe not all QKeras optmizers belong here?
-    register_flow(
-        'optimize',
-        [
-            'eliminate_linear_activation',
-            'fuse_consecutive_batch_normalization',
-            'fuse_batch_normalization',
-            'replace_multidimensional_dense_with_conv',
-            'set_precision_concat',
-        ],
-        requires=['convert'],
-    )
-except ImportError:
-    register_flow('convert', ['fuse_bias_add', 'remove_useless_transpose'])
-    register_flow(
-        'optimize',
-        [
-            'eliminate_linear_activation',
-            'fuse_batch_normalization',
-            'replace_multidimensional_dense_with_conv',
-            'set_precision_concat',
-        ],
-        requires=['convert'],
-    )
-
 del opt_path
 del module_path
 del optimizers
+
+register_flow(
+    'convert',
+    [
+        'fuse_bias_add',
+        'remove_useless_transpose',
+        'output_rounding_saturation_mode',
+        'qkeras_factorize_alpha',
+        'extract_ternary_threshold',
+        'fuse_consecutive_batch_normalization',
+    ],
+)  # TODO Maybe not all QKeras optmizers belong here?
+
+register_flow(
+    'optimize',
+    [
+        'eliminate_linear_activation',
+        'fuse_consecutive_batch_normalization',
+        'fuse_batch_normalization',
+        'replace_multidimensional_dense_with_conv',
+        'set_precision_concat',
+    ],
+    requires=['convert'],
+)
