+++ hls4ml/utils/config.py
@@ -1,5 +1,7 @@
 import json
 
+import qkeras
+
 import hls4ml
 
 
@@ -23,7 +25,6 @@ def create_config(output_dir='my-hls-test', project_name='myproject', backend='V
 
 
 def _get_precision_from_quantizer(quantizer):
-    import qkeras
 
     if isinstance(quantizer, str):
         quantizer_obj = qkeras.get_quantizer(quantizer)
